Comman Palette Category more polished, and refactoring in the other files
