+++ package.json
@@ -164,7 +164,7 @@
 				"title": "Open Collection"
 			},
 			{
-				"category": "DocDB",
+				"category": "DocumentDB",
 				"command": "cosmosDB.openDocDBCollection",
 				"title": "Open Collection"
 			}
